+++ packages/zwave-js/src/lib/driver/Driver.test.ts
@@ -41,6 +41,9 @@ const PORT_ADDRESS = "/tty/FAKE";
 
 async function createAndStartDriver() {
 	const driver = new Driver(PORT_ADDRESS, { skipInterview: true });
+	driver.on("error", () => {
+		/* swallow error events during testing */
+	});
 	await driver.start();
 	const portInstance = MockSerialPort.getInstance(PORT_ADDRESS)!;
 
@@ -75,6 +78,8 @@ describe("lib/driver/Driver => ", () => {
 	describe("starting it => ", () => {
 		it("should open a new serialport", async () => {
 			const driver = new Driver(PORT_ADDRESS, { skipInterview: true });
+			// swallow error events during testing
+			driver.on("error", () => {});
 			// start the driver
 			await driver.start();
 
@@ -85,6 +90,8 @@ describe("lib/driver/Driver => ", () => {
 
 		it("should only work once", async () => {
 			const driver = new Driver(PORT_ADDRESS, { skipInterview: true });
+			// swallow error events during testing
+			driver.on("error", () => {});
 			// start the driver twice
 			await driver.start();
 			await driver.start();
@@ -96,6 +103,8 @@ describe("lib/driver/Driver => ", () => {
 
 		it("the start promise should only be fulfilled after the port was opened", async () => {
 			const driver = new Driver(PORT_ADDRESS, { skipInterview: true });
+			// swallow error events during testing
+			driver.on("error", () => {});
 
 			// start the driver
 			const fulfilledSpy = jest.fn();
@@ -108,6 +117,8 @@ describe("lib/driver/Driver => ", () => {
 
 		it("the start promise should be rejected if the port opening fails", async () => {
 			const driver = new Driver(PORT_ADDRESS, { skipInterview: true });
+			// swallow error events during testing
+			driver.on("error", () => {});
 
 			// start the driver
 			const startPromise = driver.start();
@@ -122,6 +133,8 @@ describe("lib/driver/Driver => ", () => {
 
 		it("after a failed start, starting again should not be possible", async () => {
 			const driver = new Driver(PORT_ADDRESS, { skipInterview: true });
+			// swallow error events during testing
+			driver.on("error", () => {});
 
 			// start the driver
 			const startPromise = driver.start();
@@ -138,11 +151,21 @@ describe("lib/driver/Driver => ", () => {
 
 			driver.destroy();
 		});
+
+		it(`should throw if no "error" handler is attached`, async () => {
+			const driver = new Driver(PORT_ADDRESS, { skipInterview: true });
+			// start the driver
+			await assertZWaveError(() => driver.start(), {
+				errorCode: ZWaveErrorCodes.Driver_NoErrorHandler,
+			});
+		});
 	});
 
 	describe("sending messages => ", () => {
 		it("should not be possible if the driver wasn't started", async () => {
 			const driver = new Driver(PORT_ADDRESS, { skipInterview: true });
+			// swallow error events during testing
+			driver.on("error", () => {});
 
 			const msg = new TestMessage(driver);
 			await assertZWaveError(() => driver.sendMessage(msg), {
@@ -154,6 +177,8 @@ describe("lib/driver/Driver => ", () => {
 
 		it("should not be possible if the driver hasn't completed starting", async () => {
 			const driver = new Driver(PORT_ADDRESS, { skipInterview: true });
+			// swallow error events during testing
+			driver.on("error", () => {});
 
 			// start the driver, but don't open the serial port yet
 			driver.start();
@@ -168,6 +193,8 @@ describe("lib/driver/Driver => ", () => {
 
 		it("should not be possible if the driver failed to start", async () => {
 			const driver = new Driver(PORT_ADDRESS, { skipInterview: true });
+			// swallow error events during testing
+			driver.on("error", () => {});
 
 			// start the driver
 			const startPromise = driver.start();
@@ -297,9 +324,6 @@ describe("lib/driver/Driver => ", () => {
 		});
 
 		it("should not crash if a message is received that cannot be deserialized", () => {
-			// swallow the error
-			driver.on("error", () => {});
-
 			const req = new ApplicationCommandRequest(driver, {
 				command: new WakeUpCCIntervalSet(driver, {
 					nodeId: 1,